rewrite MapObjectsStore v0.8   - use GuidSet for make object-safe UpdateDataMapType (this fix at least pet action bar after mount)   - revert some changes from versions 3-6 (need more work)   - use unordered_map instead of TypeUnorderedMapContainer template collection   - full rewrite object update making system, make dependent update queue for players items   - correct Camera appending to not-in-world objects   - add Transport objects to current Map storing/updating system (currently by half-hack)   - simplify code NEED TEST and measurements CPU load and resource leaks! related to #1165, #1083, #852

+++ src/game/Unit.cpp
@@ -308,7 +308,7 @@ Unit::Unit() :
 Unit::~Unit()
 {
     if (IsInWorld())
-        Object::RemoveFromWorld();
+        WorldObject::RemoveFromWorld(true);
 
     ResetMap();
 
@@ -11215,11 +11215,11 @@ uint32 Unit::GetCreatePowers( Powers power ) const
 
 void Unit::AddToWorld()
 {
-    Object::AddToWorld();
+    WorldObject::AddToWorld();
     ScheduleAINotify(0);
 }
 
-void Unit::RemoveFromWorld()
+void Unit::RemoveFromWorld(bool remove)
 {
     // cleanup
     if (IsInWorld())
@@ -11238,11 +11238,14 @@ void Unit::RemoveFromWorld()
         GetViewPoint().Event_RemovedFromWorld();
     }
 
-    Object::RemoveFromWorld();
+    WorldObject::RemoveFromWorld(remove);
 }
 
 void Unit::CleanupsBeforeDelete()
 {
+    if (!IsInWorld())
+        return;
+
     if (m_uint32Values)                                      // only for fully created object
     {
         if (GetVehicle())
@@ -11251,11 +11254,11 @@ void Unit::CleanupsBeforeDelete()
             RemoveVehicleKit();
         InterruptNonMeleeSpells(true);
         KillAllEvents(false);                      // non-delatable (currently casted spells) will not deleted now but it will deleted at call in Map::RemoveAllObjectsInRemoveList
-        if (IsInWorld())
-            CombatStop();
+        CombatStop();
         ClearComboPointHolders();
-        DeleteThreatList();
-        if (GetTypeId()==TYPEID_PLAYER)
+        if (CanHaveThreatList())
+            DeleteThreatList();
+        if (GetTypeId() == TYPEID_PLAYER)
             getHostileRefManager().setOnlineOfflineState(false);
         else
             getHostileRefManager().deleteReferences();
